requirements.txt
alpaca-trade-api==0.48
pandas>=1.3.0
numpy>=1.21.0
yfinance==1.2.0
backtrader==1.9.78.123
peewee==4.0.2
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
